requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.29.0
jinja2==3.1.3
deep-translator==1.11.4
apscheduler==3.10.4
